+++ src/common/utility/vga2ansi.cpp
@@ -21,6 +21,7 @@
 **
 */
 
+#include <cstdbool>
 #include <cstdint>
 #include <cstdio>
 #include <cstdlib>
@@ -35,7 +36,7 @@ static void CPrint(const char* in)
 	DWORD bytes_written;
 	if (!StdOut)
 		return;
-	WriteFile(StdOut, in, static_cast<int>(strlen(in)), &bytes_written, NULL);
+	WriteFile(StdOut, in, strlen(in), &bytes_written, NULL);
 }
 #else
 static void CPrint(const char* in)
